# api/openapi/openapi.yaml
openapi: 3.1.0
info:
  title: Omnir CRM API
  version: 0.1.0
  description: |
    REST API for Omnir CRM — the modern replacement for vtiger CE.
    Built in Go, PostgreSQL-backed, JWT-authenticated.
  contact:
    name: Omnir Engineering
  license:
    name: Proprietary
servers:
  - url: http://localhost:8080/api/v1
    description: Local dev
  - url: https://api.omnir.io/api/v1
    description: Production
security:
  - bearerAuth: []
...
paths:
...
  /contacts:
    get:
      tags: [contacts]
      summary: List contacts
      parameters:
        - $ref: '#/components/parameters/Page'
        - $ref: '#/components/parameters/Limit'
        - $ref: '#/components/parameters/Q'
        - $ref: '#/components/parameters/OwnerId'
        - name: stage
          in: query
          schema:
            $ref: '#/components/schemas/ContactStage'
        - name: account_id
          in: query
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Paginated contacts
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ContactList'
    post:
      tags: [contacts]
      summary: Create contact
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ContactCreate'
      responses:
        '201':
          description: Created contact
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Contact'
        '422':
          $ref: '#/components/responses/UnprocessableEntity'
  /contacts/{id}:
    parameters:
      - $ref: '#/components/parameters/Id'
      - $ref: '#/components/parameters/LinkedPage'
      - $ref: '#/components/parameters/LinkedLimit'
      - $ref: '#/components/parameters/LinkedType'
      - $ref: '#/components/parameters/LinkedRole'
      - $ref: '#/components/parameters/LinkedSince'
    get:
      tags: [contacts]
      summary: Get contact
      responses:
        '200':
          description: Contact
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ContactDetail'
        '404':
          $ref: '#/components/responses/NotFound'
    patch:
      tags: [contacts]
      summary: Update contact
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ContactPatch'
      responses:
        '200':
          description: Updated contact
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Contact'
    delete:
      tags: [contacts]
      summary: Delete contact
      responses:
        '204':
          description: Deleted
...
  /accounts/{id}/contacts:
    parameters:
      - $ref: '#/components/parameters/Id'
    get:
      tags: [accounts]
      summary: List contacts for an account
      responses:
        '200':
          description: Contact list
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ContactList'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
    Id:
      name: id
      in: path
      required: true
      schema:
        type: string
        format: uuid
    Page:
      name: page
      in: query
      schema:
        type: integer
        minimum: 1
        default: 1
    Limit:
      name: limit
      in: query
      schema:
        type: integer
        minimum: 1
        maximum: 200
        default: 50
    Q:
      name: q
      in: query
      schema:
        type: string
...
    OwnerId:
      name: owner_id
      in: query
      schema:
        type: string
        format: uuid
    LinkedPage:
      name: linked_page
      in: query
      schema:
        type: integer
        minimum: 1
        default: 1
    LinkedLimit:
      name: linked_limit
      in: query
      schema:
        type: integer
        minimum: 1
        maximum: 200
        default: 25
    LinkedType:
      name: type
      in: query
      schema:
        type: string
    LinkedRole:
      name: role
      in: query
      schema:
        type: string
    LinkedSince:
      name: since
      in: query
      schema:
        type: string
        format: date-time
  responses:
...
    NotFound:
      description: Not found
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/Problem'
    UnprocessableEntity:
      description: Validation error
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/Problem'
  schemas:
    Problem:
      type: object
      description: RFC 7807 Problem Details
      properties:
        type:
          type: string
          format: uri
        title:
          type: string
        status:
          type: integer
        detail:
          type: string
        instance:
          type: string
        errors:
          type: object
          additionalProperties:
            type: array
            items:
              type: string

    # ── Module Configuration ───────────────────────────────────────────────
...
    ContactStage:
      type: string
      enum: [lead, prospect, customer, churned]
    Contact:
      type: object
      properties:
        id:
          type: string
          format: uuid
        first_name:
          type: string
        last_name:
          type: string
        email:
          type: string
          format: email
          nullable: true
        phone:
          type: string
          nullable: true
        account_id:
          type: string
          format: uuid
          nullable: true
        owner_id:
          type: string
          format: uuid
        lead_source:
          type: string
          nullable: true
        stage:
          $ref: '#/components/schemas/ContactStage'
        tags:
          type: array
          items:
            type: string
        custom_fields:
          type: object
          additionalProperties: true
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
    ContactCreate:
      type: object
      required: [first_name, last_name]
      properties:
        first_name:
          type: string
        last_name:
          type: string
        email:
          type: string
          format: email
        phone:
          type: string
        account_id:
          type: string
          format: uuid
        owner_id:
          type: string
          format: uuid
        lead_source:
          type: string
        stage:
          $ref: '#/components/schemas/ContactStage'
        tags:
          type: array
          items:
            type: string
        custom_fields:
          type: object
          additionalProperties: true
    ContactPatch:
      type: object
      properties:
        first_name:
          type: string
        last_name:
          type: string
        email:
          type: string
          format: email
          nullable: true
        phone:
          type: string
          nullable: true
        account_id:
          type: string
          format: uuid
          nullable: true
        owner_id:
          type: string
          format: uuid
        lead_source:
          type: string
          nullable: true
        stage:
          $ref: '#/components/schemas/ContactStage'
        tags:
          type: array
          items:
            type: string
        custom_fields:
          type: object
          additionalProperties: true
    ContactList:
      allOf:
        - $ref: '#/components/schemas/Pagination'
        - type: object
          properties:
            data:
              type: array
              items:
                $ref: '#/components/schemas/Contact'
    LinkedEntity:
      type: object
      required: [type, id, role, source, created_at]
      properties:
        type:
          type: string
        id:
          type: string
          format: uuid
        role:
          type: string
        source:
          type: string
        created_at:
          type: string
          format: date-time
    LinkedEntitiesMeta:
      type: object
      properties:
        total:
          type: integer
        page:
          type: integer
        limit:
          type: integer
    ContactDetail:
      allOf:
        - $ref: '#/components/schemas/Contact'
        - type: object
          properties:
            linked_entities:
              type: array
              items:
                $ref: '#/components/schemas/LinkedEntity'
            associations:
              description: Deprecated compatibility alias for linked_entities.
              type: array
              items:
                $ref: '#/components/schemas/LinkedEntity'
            linked_entities_meta:
              $ref: '#/components/schemas/LinkedEntitiesMeta'

    # ── Account ─────────────────────────────────────────────────────────────
...
    Pagination:
      type: object
      properties:
        total:
          type: integer
        page:
          type: integer
        limit:
          type: integer
        total_pages:
          type: integer